=> luna/projects.py            - Created a new list called '_paths' to store the names of all paths required in a project. Based on that list, the algorithm now which paths to create and remove.            - Added the new function remove_empty_paths() to remove paths that are empty.            - Added code to verify which entries failed. IFPs and MFPs will be created only if at least one entry finished successfully.            - I returned setrecursionlimit() to the main code again due to the RecursionError raised by pickle. This error occurs due to the circular references in the results data structure. In the future, I plan to remove those circular references to avoid those errors.
=> luna/util/default_values.py:
          - Added a new default variable called RECURSION_LIMIT, which is used in project.py to temporarily fix the RecursionError due to the pickle library. 

=> luna/util/file.py:
         - Added a new parameter called 'only_empty_paths' to clear_directory(). This parameter controls if it should remove only empty paths or all paths.

=> luna/MyBio/PDB/PDBIO:
         - Fixed problem with the addition of CONECTS to output files when preserve_atom_numbering is set to False. In this situation, the serial numbers are changed and the CONECT reference of the atoms is not accessed anymore as the key to access it is the old serial number. Now it creates a mapping to save the old serial numbers.

+++ luna/util/file.py
@@ -1,5 +1,5 @@
 from os.path import basename, exists, isdir, isfile, splitext
-from os import makedirs, remove
+from os import makedirs, remove, listdir
 from shutil import rmtree
 import string
 import random
@@ -110,11 +110,15 @@ def remove_files(files):
         if exists(f):
             remove(f)
         else:
-            logger.info("File %s does not exist." % f)
+            logger.info("File '%s' does not exist." % f)
 
 
-def clear_directory(path):
+def clear_directory(path, only_empty_paths=False):
     if isdir(path):
+        # Do nothing if the directory is not empty and if only empty paths must be removed.
+        if only_empty_paths and len(listdir(path)) != 0:
+            return
+
         try:
             rmtree(path, ignore_errors=True)
         except OSError as e:
